backend/requirements.txt
django==5.1.7
djangorestframework
django-cors-headers
gunicorn
whitenoise
yfinance
pyportfolioopt
numpy
pandas
scipy